Completed HUD
I added some nice fonts to give it an eerie look. Repositioned it to the
bottom left corner, because I believe it belongs there. I set the bars
to a minimum of 5 pixels so they don't disappear to keep the design
pretty. Prevented the HUD from drawing if the player is dead. I can call
this completed, but I may add weapon and ammo support in the future.

Here is an album of screenshots: http://imgur.com/a/gnlxf