Basic Ammo Counter
Created a basic ammo counter on the left side. I am not using
LocalPlayer():Clip1() yet, since we do not have weapons ready. The text
for the ammo is aligned to the right, so the ammo can go as high as the
panel's width.

Imgur image: http://imgur.com/7WmA1kt