Created basic team selection panel
Why does it only cover half of the screen? You will see.

I have to fix the text on the right half of the panel, which I will do
once I get back home.

The plan:
Get all of the teams in a list on the left half of the screen. The when
one of the teams is selected, a panel on the right side will show a
model of the player, his loadout/perks, and other information we want to
include.